requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pytest==8.3.3
httpx==0.27.2
python-telegram-bot==21.6
requests==2.32.3
